requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
attrs==22.1.0
certifi==2022.9.24
chardet==3.0.4
charset-normalizer==2.1.1
colorama==0.4.6
colorlog==6.7.0
docopt==0.6.2
frozenlist==1.3.3
googletrans-py==4.0.0
h11==0.14.0
h2==4.1.0
hikari==2.0.0.dev113
hikari-lightbulb==2.3.1
hikari-miru==2.0.1
hpack==4.0.0
hstspreload==2022.12.1
httpcore==0.16.2
httpx==0.23.1
hyperframe==6.0.1
idna==3.4
multidict==6.0.3
Pillow==9.3.0
requests==2.28.1
rfc3986==1.5.0
sniffio==1.3.0
urllib3==1.26.13
yarg==0.1.9
yarl==1.8.2
